% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-package.R
\docType{package}
\name{clock-package}
\alias{clock-package}
\alias{clock}
\title{Time Classes}
\description{
Time Classes
}
\author{
Scott Chamberlain \href{mailto:sckott@protonmail.com}{sckott@protonmail.com}
}
\keyword{package}
