% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\docType{data}
\name{Clock}
\alias{Clock}
\alias{clock_now}
\title{Clock class}
\usage{
clock_now(tzone = "", ...)
}
\arguments{
\item{tzone}{(character) a valid time zone, defaults to your local timezone}

\item{...}{options to initializer}
}
\value{
a \link{POSIXct} object
}
\description{
Clock class
}
\details{
\strong{Methods}
\describe{
\item{\code{now(tzone = "")}}{
get the date time right now
param \code{tzone}: a valid time zone
}
}
}
\examples{
x <- Clock$new()
x
x$time
x$now()
x$utc()
x$now("UTC")

Clock$new(2009)$time
Clock$new(2009, 3, 13)$time
Clock$new(2009, 3, 13, 1, 4, 53)$time

clock_now()
clock_now("UTC")
}
\keyword{datasets}
