% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_mock.R
\name{clock_mock}
\alias{clock_mock}
\title{Mock time}
\usage{
clock_mock(on = TRUE)
}
\arguments{
\item{on}{(logical) turn mocking on with \code{TRUE} or turn off with \code{FALSE}.
By default is \code{FALSE}}
}
\description{
Mock time
}
\details{
\code{timefuzz} package required for mocking behavior
}
\examples{
\dontrun{

if (interactive()) {
  # load timefuzz
  library(timefuzz)
  library(clock)

  # turn on mocking
  clock_mock()

  # do stuff
  x <- time_fuzz$new()
  ## set to today + 435 days
  x$freeze(Sys.Date() + 435)

  # get time again
  now()

  # turn off mocking
  clock_mock(FALSE)

  # get time again
  now()
}

}
}
