% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\name{clock_now}
\alias{clock_now}
\title{now}
\usage{
clock_now(tzone = "", ...)
}
\arguments{
\item{tzone}{(character) a valid time zone, defaults to your local timezone}

\item{...}{options to initializer}
}
\description{
now
}
