% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Now.R
\name{now}
\alias{now}
\title{now}
\usage{
now(tzone = "")
}
\arguments{
\item{tzone}{(character) a valid time zone, defaults to your local timezone}
}
\description{
now
}
\examples{
x <- now()
x
x$time
x$date
x$utc
x$unix_epoch
x$min
x$sec

# mocking
Sys.Date()
clock_mock()
library(timefuzz)
mock <- time_fuzz$new()
## set to today + 10 days
mock$freeze(Sys.Date() + 10)
z <- now()
z
z$time
z$date
z$utc
z$unix_epoch
z$min
z$sec
}
