% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_date.R
\name{sys_date}
\alias{sys_date}
\title{Mockable \code{\link[=Sys.Date]{Sys.Date()}} drop in replacement}
\usage{
sys_date()
}
\value{
\code{Date}, matching what \code{\link[=Sys.Date]{Sys.Date()}} returns
}
\description{
Mockable \code{\link[=Sys.Date]{Sys.Date()}} drop in replacement
}
\examples{
Sys.Date()
sys_date()

clock_mock()
library(timefuzz)
x <- time_fuzz$new()
x$freeze(Sys.Date() + 5)
sys_date()
clock_mock(FALSE)
sys_date()
}
\seealso{
\code{\link[=sys_time]{sys_time()}}
}
