% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_time.R
\name{sys_time}
\alias{sys_time}
\title{Mockable \code{\link[=Sys.time]{Sys.time()}} drop in replacement}
\usage{
sys_time()
}
\value{
\code{POSIXct}, matching what \code{\link[=Sys.time]{Sys.time()}} returns
}
\description{
Mockable \code{\link[=Sys.time]{Sys.time()}} drop in replacement
}
\examples{
Sys.time()
sys_time()

clock_mock()
library(timefuzz)
x <- time_fuzz$new()
x$freeze(Sys.time() + 500)
sys_time()
clock_mock(FALSE)
sys_time()
}
\seealso{
\code{\link[=sys_date]{sys_date()}}
}
