% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\name{clock}
\alias{clock}
\title{Clock class}
\arguments{
\item{...}{options to initializer}

\item{tzone}{(character) a valid time zone, defaults to your local timezone}
}
\value{
a \link{POSIXct} object
}
\description{
Clock class
}
\details{
\strong{Methods}
\describe{
\item{\code{now(tzone = "")}}{
get the date time right now
param \code{tzone}: a valid time zone
}
}
}
\examples{
x <- clock()
x
x$time
x$now()
x$utc()
x$now("UTC")
x$date()

clock(2009)$time
clock(2009, 3, 13)$time
clock(2009, 3, 13, 1, 4, 53)$time

clock_now()
clock_now("UTC")
}
